Uncaught Error: Cannot find module 'jquery-ui/autocomplete' #114
Fix issue #114 by adding a dependency on jquery-ui. This requires to pin
an older version of jquery-ui (1.10.5) since this matches the currently
committed version.

For some versions Atom seems to override the existing node_modules so
that's why the checked in version does not seem to be working in some
cases.